+++ src/SmsManager.php
@@ -8,7 +8,9 @@ use AliYavari\IranSms\Abstracts\Driver;
 use AliYavari\IranSms\Contracts\Sms;
 use AliYavari\IranSms\Drivers\AmootSmsDriver;
 use AliYavari\IranSms\Drivers\FakeDriver;
+use AliYavari\IranSms\Drivers\FaraPayamakDriver;
 use AliYavari\IranSms\Drivers\FarazSmsDriver;
+use AliYavari\IranSms\Drivers\GhasedakDriver;
 use AliYavari\IranSms\Drivers\KavenegarDriver;
 use AliYavari\IranSms\Drivers\MeliPayamakDriver;
 use AliYavari\IranSms\Drivers\PayamResanDriver;
@@ -105,6 +107,16 @@ final class SmsManager extends Manager
     protected function createAmootSmsDriver(): AmootSmsDriver
     {
         return $this->container->make(AmootSmsDriver::class);
+    }
+
+    protected function createFaraPayamakDriver(): FaraPayamakDriver
+    {
+        return $this->container->make(FaraPayamakDriver::class);
+    }
+
+    protected function createGhasedakDriver(): GhasedakDriver
+    {
+        return $this->container->make(GhasedakDriver::class);
     }
 
     private function mustBeFresh(string $driver): bool