+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Added
+
+- **Fara Payamak** provider.
+- **Ghasedak** provider.
+
 ## [1.1.1] - 2025-08-12
 
 ### Fixed
@@ -41,6 +48,7 @@ Supported Providers:
 - Faraz SMS
 - Raygan SMS
 
+[Unreleased]: https://github.com/amyavari/iran-sms-laravel/compare/v1.1.1...HEAD
 [1.1.1]: https://github.com/amyavari/iran-sms-laravel/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/amyavari/iran-sms-laravel/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/amyavari/iran-sms-laravel/compare/v0.1.0...v1.0.0